processor/requirements.txt
Flask
connexion
mysql-connector
mysql-connector-python
pymongo
